aligning generator content with main branch
