+++ src/components/ccsds_command_depacketizer/gen/templates/assembly/name_ccsds_cosmos_commands.txt
@@ -1,32 +1,37 @@
 {% for id, command in commands.items() %}
-COMMAND {{ name|upper }} {{ command.suite.component.instance_name|upper }}-{{ command.name|upper }} BIG_ENDIAN "{{ command.description }}"
+Command {{ name }} {{ command.suite.component.instance_name }}-{{ command.name }} Big_Endian "{{ command.description }}"
 {# The ccsds command header #}
 {# This header assumed by the Ccsds_Command_Depacketizer component #}
-  PARAMETER CCSDSVER 0 3 UINT 0 7 0 "Packet version number"
-  PARAMETER CCSDSTYPE 3 1 UINT 0 1 1 "Packet type"
-    STATE TLM 0
-    STATE CMD 1
-  PARAMETER CCSDSSHF 4 1 UINT 0 1 1 "CCSDS secondary header flag"
-    STATE FALSE 0
-    STATE TRUE 1
-  PARAMETER CCSDSAPID 5 11 UINT 0 2047 8 "CCSDS application process ID"
-  PARAMETER CCSDSSEQFLAGS 16 2 UINT 0 3 3 "Sequence flags"
-    STATE FIRST 0
-    STATE CONT 1
-    STATE LAST 2
-    STATE NOGROUP 3
-  PARAMETER CCSDSSEQNT 18 14 UINT MIN MAX 0 "Packet sequence count"
-  PARAMETER CCSDSLENGTH 32 16 UINT MIN MAX {% if command.size != None %}{{ (command.type_model.fields|length * (command.size / 8) + 3)|int }}{% else %}3{% endif %} "Packet data length"
-  PARAMETER CCSDSSHDRRB 48 1 UINT MIN MAX 0 "Secondary header reserve bit"
-  PARAMETER CCSDSSHDRFC 49 7 UINT MIN MAX 0 "Secondary header function code"
-  PARAMETER CCSDSSHDRCRC 56 8 UINT MIN MAX 0 "Secondary header checksum"
-  ID_PARAMETER CCSDSCMDID 64 16 UINT MIN MAX {{ command.id }} "Adamant command ID"
+  Parameter Ccsdsver 0 3 UINT 0 7 0 "Packet version number"
+  Parameter Ccsdstype 3 1 UINT 0 1 1 "Packet type"
+    State TLM 0
+    State CMD 1
+  Parameter Ccsdsshf 4 1 UINT 0 1 1 "CCSDS secondary header flag"
+    State FALSE 0
+    State TRUE 1
+  Parameter Ccsdsapid 5 11 UINT 0 2047 8 "CCSDS application process ID"
+  Parameter Ccsdsseqflags 16 2 UINT 0 3 3 "Sequence flags"
+    State FIRST 0
+    State CONT 1
+    State LAST 2
+    State NOGROUP 3
+  Parameter Ccsdssqnt 18 14 UINT MIN MAX 0 "Packet sequence count"
+  Parameter Ccsdslength 32 16 UINT MIN MAX {% if command.size != None %}{{ (command.type_model.fields|length * (command.size / 8) + 3)|int }}{% else %}3{% endif %} "Packet data length"
+  Parameter Ccsdsshdrrb 48 1 UINT MIN MAX 0 "Secondary header reserve bit"
+  Parameter Ccsdsshdrfc 49 7 UINT MIN MAX 0 "Secondary header function code"
+  Parameter Ccsdsshdrcrc 56 8 UINT MIN MAX 0 "Secondary header checksum"
+  ID_Parameter Ccsdscmdid 64 16 UINT MIN MAX {{ command.id }} "Adamant command ID"
 {# The command argument definitions. #}
-{% if command.type %}{% if command.type_model %}{% for field in command.type_model.fields %}
-  APPEND_PARAMETER {{ field|upper }} {{ command.size -}}
-{%- if 'F32' in command.type %} FLOAT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}
+{% if command.type %}{% if command.type_model %}{% for field_name, field in command.type_model.fields.items() %}
+  Append_Parameter {% if field.type_model.name %}{{ field.type_model.name }}.{% endif %}{{ field.name }} {% if field.format.type %}{{ field.format.type[1:] }}{% endif %}{% if field.type_model.size %}{{ field.type_model.size }}{% endif -%}
+{% if field.format.type %}{%- if 'F' in field.format.type[0] %} FLOAT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}
+{%- elif 'U' in field.format.type[0] %} UINT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}
+{%- elif 'E' in field.format.type[0] %} UINT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}
+{%- elif 'I' in field.format.type[0] %} INT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}{% endif %}
 {%- elif 'Buffer' in command.type or command.size > 64 %} BLOCK 0
-{%- else %} UINT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}{% endif %} "{{ command.type_model.description|replace('\n','') }}"
-{% endfor %}{% endif %}{% endif %}
+{%- else %} *UINT MIN MAX {% if command.default_value %}{{ command.default_value }}{% else %}0{% endif %}{% endif %} "{{ field.description|replace('\n','') }}"
+{% if field.format.type and 'E' in field.format.type[0] %}{% for literals in field.type_model.literals %}
+    State {{ literals.name }} {{ literals.value }}
+{% endfor %}{% endif %}{% endfor %}{% endif %}{% endif %}
 
 {% endfor %}
