--- src/components/ccsds_command_depacketizer/gen/templates/assembly/name_ccsds_cosmos_plugin.txt
@@ -1,11 +1,0 @@
-# Set VARIABLEs here to allow variation in your plugin
-VARIABLE {{ name|lower }}_target_name {{ name|upper }}
-VARIABLE port_w /dev/ttyACM0
-VARIABLE port_r /dev/ttyACM0
-VARIABLE checksum_parameter_name CCSDSSHDRCRC
-
-TARGET {{ name|upper }} <%= {{ name|lower }}_target_name %>
-INTERFACE <%= {{ name|lower }}_target_name %>_INT serial_interface.rb <%= port_w %> <%= port_r %> 10.0 nil LENGTH 64 16 11 1 BIG_ENDIAN 0 FED4AFEE
-  MAP_TARGET <%= {{ name|lower }}_target_name %>
-  PROTOCOL READ crc_sync_protocol.rb nil true "ERROR" -16 16
-  PROTOCOL WRITE cmd_sync_checksum.rb <%= checksum_parameter_name %>
