+++ src/components/ccsds_packetizer/gen/generators/ccsds_xml.py
@@ -3,7 +3,7 @@ import models.assembly
 from generators import assembly as genassem
 from base_classes.generator_base import generator_base
 from util import model_loader
-from generators.assembly import assembly_hydra_generator
+from generators.assembly import assembly_hydra_generator, assembly_cosmos_plugin_generator_base
 
 
 class ccsds_packets_xml(assembly_hydra_generator, generator_base):
@@ -35,5 +35,19 @@ class ccsds_packets_xml(assembly_hydra_generator, generator_base):
 
         a.type_format_dictionary = genassem.type_format_dictionary
         genassem.create_type_print_strings(a)
+        genassem.create_type_field_strings(a)
+        print(a.render(self.template, template_path=self.template_dir))
+
+
+class assembly_cosmos_telemetry_txt(assembly_cosmos_plugin_generator_base, generator_base):
+    def __init__(self):
+        this_file_dir = os.path.dirname(os.path.realpath(__file__))
+        template_dir = os.path.join(this_file_dir, ".." + os.sep + "templates")
+        assembly_cosmos_plugin_generator_base.__init__(
+            self, "name_ccsds_cosmos_telemetry.txt", template_dir=template_dir, subdir="plugin"
+        )
+
+    def generate(self, input_filename):
+        a = self.model_cls(input_filename)
         genassem.create_type_field_strings(a)
         print(a.render(self.template, template_path=self.template_dir))
