+++ src/components/ccsds_packetizer/gen/templates/assembly/name_ccsds_cosmos_telemetry.txt
@@ -1,0 +1,39 @@
+{% for id, packet in packets.items() %}
+Telemetry {{ name }} {{ packet.name }} Big_Endian "{{ packet.description }}"
+{# The ccsds packets header #}
+{# This header assumed by the Ccsds_packets_Depacketizer component #}
+  Item Ccsdsver 0 3 UINT "CCSDS packet version number"
+  Item Ccsdstype 3 1 UINT "CCSDS packet type (Command or Telemetry)"
+    State TLM 0
+    State CMD 1
+  Item Ccsdsshf 4 1 UINT "CCSDS secondary header flag"
+    State FALSE 0
+    State TRUE 1
+  ID_Item Ccsdsapid 5 11 UINT {{ packet.id }} "CCSDS application process ID"
+  Item Ccsdsseqflags 16 2 UINT "CCSDS sequence flags"
+    State FIRST 0
+    State CONT 1
+    State LAST 2
+    State NOGROUP 3
+  Item Ccsdssqnt 18 14 UINT "CCSDS packet sequence count"
+  Item Ccsdslength 32 16 UINT "CCSDS packet data length"
+  Item Ccsdssec 48 32 UINT "Seconds since epoch"
+  Item Ccsdssubsec 80 32 UINT "Milliseconds of day (0 - 86399999)"
+{# The packets definitions. #}
+{% for item_name, item in packet.items.items() %}
+  Append_Item {{ item.entity.full_name }}{% if item.flat_name %}.{{ item.flat_name }}{% endif %} {{ item.size -}}
+{%- if 'F' in item.format.type[0] %} FLOAT "
+{%- elif 'U' in item.format.type[0] %} UINT "
+{%- elif 'E' in item.format.type[0] %} UINT "
+{%- elif 'I' in item.format.type[0] %} INT "
+{%- elif 'Buffer' in item.type or item.size > 64 %} BLOCK "
+{%- else %} UINT "{%- endif -%}
+{{ item.flat_desc|replace('\n','') }}"
+{% if item.is_enum %}
+{% for literals in item.datatype.model.literals %}
+    State {{ literals.name }} {{ literals.value }}
+{% endfor %}{% endif %}
+{% endfor %}
+  Append_Item Subpacket.Data 0 BLOCK "Subpacket data"
+
+{% endfor %}
